牛乳選択シーンコンポーネント作成
